mile1/requirements.txt
dash
dash_bootstrap_components
pandas
plotly.express
gunicorn